Read Hobo CSVs

Read Hobo CSVs

read_hobo_csv(
  file = ".",
  orders = c("Ymd HMS", "dmy HMS"),
  units = "degC",
  tz = "Etc/GMT+8",
  recursive = FALSE,
  quiet = FALSE
)

Arguments

file
A string of the file or directory.

orders
A character vector of date-time formats used by parse_date_time.

units
A string of the units to convert the temperature data to using ud.convert.

tz
A string of the time zone to convert the date times to using with_tz.

recursive
A flag indicating whether to read files from subdirectories.

quiet
A flag indicating whether to provide messages. Ignored if file is a file (as opposed to a directory).

Value

A tibble of the data with the temperature.

Examples

read_hobo_csv(system.file("hobo", "10723440.csv", package = "poisdata"))

#> Warning: `filter_()` is deprecated as of dplyr 0.7.0.
#> Please use `filter()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Warning: `select_()` is deprecated as of dplyr 0.7.0.
#> Please use `select()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> imported 11 rows of data from '/Users/runner/work/_temp/Library/poisdata/hobo/10723440.csv'

#> Warning: `as.tbl()` is deprecated as of dplyr 1.0.0.
#> Please use `tibble::as_tibble()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> # A tibble: 11 x 6
#>    Logger  DateTime            Temperature_degC FileRow FileName Directory      
#>    <chr>   <dttm>                         <dbl>   <dbl> <chr>    <chr>          
#>  1 107234… 2015-11-10 07:18:12            14.6        1 10723440 /Users/runner/…
#>  2 107234… 2015-11-10 07:23:12            17.1        3 10723440 /Users/runner/…
#>  3 107234… 2015-11-10 07:28:12            16.2        4 10723440 /Users/runner/…
#>  4 107234… 2015-11-10 07:33:12            13.1        5 10723440 /Users/runner/…
#>  5 107234… 2015-11-10 07:38:12            10.9        6 10723440 /Users/runner/…
#>  6 107234… 2015-11-10 07:43:12             9.44       7 10723440 /Users/runner/…
#>  7 107234… 2015-11-10 07:48:12             8.42       8 10723440 /Users/runner/…
#>  8 107234… 2015-11-10 07:53:12             7.80       9 10723440 /Users/runner/…
#>  9 107234… 2015-11-10 07:58:12             7.39      10 10723440 /Users/runner/…
#> 10 107234… 2016-04-09 04:38:12             3.93      11 10723440 /Users/runner/…
#> 11 107234… 2016-04-09 04:43:12             3.93      12 10723440 /Users/runner/…